Extend new technique to Core_Archon
The previous commit left the "old" workaround in place in Core_Archon,
as the code appeared to work, and we were interested in making fewer
changes.  However, it appears the above code branch is used exactly
once in Archon (the Collections core mixin), and the initialize
function there was simply not running.

This change gets it to run again as expected.

Signed-off-by: Dan Wells <dbw2@calvin.edu>